requirements.txt
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
coveralls==1.3.0
dj-database-url==0.5.0
Django==2.0.3
django-appconf==1.0.2
django-compressor==2.2
docopt==0.6.2
gunicorn==19.7.1
idna==2.6
libsass==0.14.2
psycopg2==2.7.4
python-dateutil==2.7.2
pytz==2018.4
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
sqlite3==0.0.0
urllib3==1.22
whitenoise==3.3.1
